+++ js/packages/trellis/sdk/activity.ts
@@ -1,1 +1,0 @@
-export * from "#trellis-generated-sdk/activity";